# Improved runtime performance by adding build-time optimization options