--- a/build_config.rb
+++ b/build_config.rb
@@ -96,7 +96,7 @@ if build_targets.include?('windows-amd64')
 
     [conf.cc, conf.linker].each do |cc|
       cc.command = "#{conf.host_target}-gcc-posix"
-      cc.flags += %w[-static -O3 -mtune=native -march=native]
+      cc.flags += %w[-static -O3]
     end
     conf.cc.defines      += %w[MRB_STR_LENGTH_MAX=0 MRB_UTF8_STRING MRUBY_YAML_NO_CANONICAL_NULL]
     conf.cxx.command      = "#{conf.host_target}-g++"