ビルド時に最適化オプションをつけることで高速化します。

```
# before
❯ time rf -g -q hello /usr/share/dict/american-english

________________________________________________________
Executed in    7.27 secs    fish           external
   usr time    6.14 secs  420.00 micros    6.14 secs
   sys time    0.80 secs    0.00 micros    0.80 secs

# after
❯ time ./build/bin/rf -q -g hello /usr/share/dict/american-english

________________________________________________________
Executed in    1.92 secs    fish           external
   usr time    1.83 secs  192.00 micros    1.83 secs
   sys time    0.00 secs  221.00 micros    0.00 secs
```